fqdn: extract dnsrequesthandler
This commit extracts the DNS request handling from the
FQDN boostrapper into its own Hive component.

Logically it's not part of the bootstrap process. By extracting,
it can also be easier reused in other components (e.g. upcoming
standalone dns proxy)
